+++ src/hooks/form/LoginForm.tsx
@@ -1,7 +1,3 @@
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
-
 import { Button } from "@/components/ui/button";
 import {
   Form,
@@ -12,34 +8,21 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import { Card, CardTitle, CardContent, CardHeader, CardDescription, CardFooter } from "@/components/ui/card";
+import {
+  Card,
+  CardTitle,
+  CardContent,
+  CardHeader,
+  CardDescription,
+  CardFooter,
+} from "@/components/ui/card";
 
 import useLogin from "../login/useLogin";
-
-const formSchema = z.object({
-  email: z.string().email({
-    message: "email형식이 아닙니다",
-  }),
-  password: z.string(),
-});
+import useLoginForm from "./useLoginForm";
 
 const LoginForm = () => {
-  const { signIn, navigate } = useLogin();
-
-  const form = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      email: "",
-      password: "",
-    },
-  });
-
-  function onSubmit(values: z.infer<typeof formSchema>) {
-    // Do something with the form values.
-    // ✅ This will be type-safe and validated.
-    signIn(values);
-    console.log(values);
-  }
+  const { navigate } = useLogin();
+  const { form, onSubmit } = useLoginForm();
 
   return (
     <>
@@ -79,9 +62,13 @@ const LoginForm = () => {
               />
               <CardFooter>
                 <div className="space-x-5">
-                  <Button onClick={()=>{
-                    navigate("/signup")
-                  }}>회원가입</Button>
+                  <Button
+                    onClick={() => {
+                      navigate("/signup");
+                    }}
+                  >
+                    회원가입
+                  </Button>
                   <Button type="submit">로그인</Button>
                 </div>
               </CardFooter>
